Extract unit tests from GeneratePropertiesIT
