+++ java/clients/src/test/java/sleeper/clients/admin/deploy/GeneratePropertiesIT.java
@@ -36,20 +36,8 @@ import java.nio.file.Path;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.configuration.properties.SystemDefinedInstanceProperty.CONFIG_BUCKET;
-import static sleeper.configuration.properties.SystemDefinedInstanceProperty.QUERY_RESULTS_BUCKET;
-import static sleeper.configuration.properties.SystemDefinedInstanceProperty.VERSION;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.ACCOUNT;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.BULK_IMPORT_REPO;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.ECR_COMPACTION_REPO;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.ECR_INGEST_REPO;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.ID;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.JARS_BUCKET;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.REGION;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.SUBNET;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.VPC_ID;
-import static sleeper.configuration.properties.table.TableProperty.DATA_BUCKET;
-import static sleeper.configuration.properties.table.TableProperty.TABLE_NAME;
 import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
 
 @Testcontainers
@@ -69,60 +57,47 @@ public class GeneratePropertiesIT {
     @Test
     void shouldGenerateInstancePropertiesCorrectly() {
         // Given/When
-        InstanceProperties properties = generateTestInstanceProperties();
+        InstanceProperties properties = generateInstancePropertiesBuilder()
+                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
+                .build().generate();
 
         // Then
-        assertThat(properties.get(ID)).isEqualTo("test-instance");
-        assertThat(properties.get(CONFIG_BUCKET)).isEqualTo("sleeper-test-instance-config");
-        assertThat(properties.get(JARS_BUCKET)).isEqualTo("sleeper-test-instance-jars");
-        assertThat(properties.get(QUERY_RESULTS_BUCKET)).isEqualTo("sleeper-test-instance-query-results");
         assertThat(properties.get(ACCOUNT)).isEqualTo(sts.getCallerIdentity(new GetCallerIdentityRequest()).getAccount());
         assertThat(properties.get(REGION)).isEqualTo(localStackContainer.getRegion());
-        assertThat(properties.get(VPC_ID)).isEqualTo("some-vpc");
-        assertThat(properties.get(SUBNET)).isEqualTo("some-subnet");
-        assertThat(properties.get(ECR_COMPACTION_REPO)).isEqualTo("test-instance/compaction-job-execution");
-        assertThat(properties.get(ECR_INGEST_REPO)).isEqualTo("test-instance/ingest");
-        assertThat(properties.get(BULK_IMPORT_REPO)).isEqualTo("test-instance/bulk-import-runner");
-        // Should not set sleeper version (system defined)
-        assertThat(properties.get(VERSION)).isNull();
-    }
-
-    @Test
-    void shouldGenerateTablePropertiesCorrectly() {
-        InstanceProperties instanceProperties = generateTestInstanceProperties();
-        TableProperties tableProperties = generateTestTableProperties(instanceProperties);
-
-        assertThat(tableProperties.get(TABLE_NAME)).isEqualTo("test-table");
-        assertThat(tableProperties.get(DATA_BUCKET)).isEqualTo("sleeper-test-instance-table-test-table");
     }
 
     @Test
     void shouldSaveBucketNamesToLocalDirectoryWhenInstancePropertiesGenerated() throws IOException {
-        InstanceProperties instanceProperties = generateTestInstanceProperties();
+        // Given
+        InstanceProperties properties = generateInstancePropertiesBuilder()
+                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
+                .build().generate();
 
-        SaveLocalProperties.saveToDirectory(tempDir, instanceProperties, Stream.empty());
+        // When
+        SaveLocalProperties.saveToDirectory(tempDir, properties, Stream.empty());
+
+        // Then
         assertThat(tempDir.resolve("configBucket.txt")).exists();
         assertThat(tempDir.resolve("queryResultsBucket.txt")).exists();
     }
 
     @Test
     void shouldSaveBucketNamesToLocalDirectoryWhenTablePropertiesGenerated() throws IOException {
-        InstanceProperties instanceProperties = generateTestInstanceProperties();
-        TableProperties tableProperties = generateTestTableProperties(instanceProperties);
+        // Given
+        InstanceProperties instanceProperties = generateInstancePropertiesBuilder()
+                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
+                .build().generate();
+        TableProperties tableProperties = GenerateTableProperties.from(instanceProperties, schemaWithKey("key"), "test-table");
 
+        // When
         SaveLocalProperties.saveToDirectory(tempDir, instanceProperties, Stream.of(tableProperties));
+
+        // Then
         assertThat(tempDir.resolve("tables/test-table/tableBucket.txt")).exists();
     }
 
-    private InstanceProperties generateTestInstanceProperties() {
+    private GenerateInstanceProperties.Builder generateInstancePropertiesBuilder() {
         return GenerateInstanceProperties.builder()
-                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
-                .instanceId("test-instance")
-                .vpcId("some-vpc").subnetId("some-subnet")
-                .build().generate();
-    }
-
-    private TableProperties generateTestTableProperties(InstanceProperties properties) {
-        return GenerateTableProperties.from(properties, schemaWithKey("key"), "test-table");
+                .instanceId("test-instance").vpcId("some-vpc").subnetId("some-subnet");
     }
 }
